+++ pygrinder/missing_completely_at_random/__init__.py
@@ -4,3 +4,9 @@
 
 # Created by Wenjie Du <wenjay.du@gmail.com>
 # License: GLP-v3
+
+from .mcar import mcar
+
+__all__ = [
+    "mcar",
+]
